refactor(services): extract ReadDotEnv/MergeDotEnv into projectenv
