+++ internal/projectenv/dotenv_test.go
@@ -1,4 +1,4 @@
-package services
+package projectenv
 
 import (
 	"os"
